backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
langgraph
langchain
python-dotenv
playwright
playwright-stealth
psycopg2-binary